requirements.txt
homeassistant==2024.2.2
voluptuous~=0.13.1
requests==2.31.0
boto3==1.34.70
python-dateutil~=2.9.0.post0
